Make some interactive functions more intelligent by selecting a project
Following functions used to work only for current project, when called
interactively:

- lab-list-project-merge-requests
- lab-list-project-pipelines
- lab-act-on-last-project-pipeline
- lab-act-on-last-failed-pipeline-job

and if the current buffer was not a part of a git project or the git
project is not a real GitLab project, then this functions would simply
fail.  Now, instead of failing, they will ask for you to select
another project interactively and work on them.